Remove sanity checks for long long types in matplotlibcpp.h
The removed checks were part of a template mechanism for type selection based on size in the matplotlibcpp.h file. These checks were asserting the sizes of long long and unsigned long long to be 8 bytes, but they have been removed to allow compatibility with different platforms with varied data type sizes.